File consistanty pass accross the Enumeration classes. This is just to get the files organized the same way and named consistant.
