+++ DataDictionary.Resource/Enumerations/DbConstraint.Type.cs
@@ -6,34 +6,5 @@ using System.Threading.Tasks;
 
 namespace DataDictionary.Resource.Enumerations
 {
-    /// <summary>
-    /// List of supported Constraint Types.
-    /// </summary>
-    public enum DbConstraintType
-    {
-        /// <summary>
-        /// Unknown Constraint Type
-        /// </summary>
-        Null,
 
-        /// <summary>
-        /// Check Constraint
-        /// </summary>
-        Check,
-
-        /// <summary>
-        /// Unique Key Constraint
-        /// </summary>
-        Unique,
-
-        /// <summary>
-        /// Primary Key Constraint
-        /// </summary>
-        PrimaryKey,
-
-        /// <summary>
-        /// Foreign Key Constraint
-        /// </summary>
-        ForeignKey,
-    }
 }
