--- DataDictionary.Resource/Enumerations/DbRoutine.Type.cs
@@ -1,23 +1,0 @@
-﻿namespace DataDictionary.Resource.Enumerations;
-
-/// <summary>
-/// List of supported Routine Types.
-/// </summary>
-public enum DbRoutineType
-{
-    /// <summary>
-    /// Unknown Routine Type
-    /// </summary>
-    Null,
-
-    /// <summary>
-    /// SQL Function
-    /// </summary>
-    Function,
-
-    /// <summary>
-    /// SQL Procedure
-    /// </summary>
-    Procedure,
-}
-
