--- DataDictionary.Resource/Enumerations/Directory.cs
@@ -1,13 +1,0 @@
-﻿// Place holder for file nesting purposes.
-namespace DataDictionary.Resource.Enumerations;
-
-/// <summary>
-/// Interface for TemplateDirectoryType
-/// </summary>
-public interface ITemplateDirectory
-{
-    /// <summary>
-    /// Root Directory to place documents in (must be a supported Special Folder).
-    /// </summary>
-    DirectoryType RootDirectory { get; }
-}
