+++ DataDictionary.Resource/Enumerations/DirectoryType.Enumeration.cs
