--- DataDictionary.Resource/Enumerations/DomainProperty.cs
@@ -1,13 +1,0 @@
-﻿// Place holder for file nesting purposes.
-namespace DataDictionary.Resource.Enumerations;
-
-/// <summary>
-/// Interface for DomainPropertyType
-/// </summary>
-public interface IDomainPropertyType
-{
-    /// <summary>
-    /// Domain Property Type for the Item.
-    /// </summary>
-    DomainPropertyType PropertyType { get; }
-}
