+++ DataDictionary.Resource/Enumerations/DomainPropertyType.Enumeration.cs
