+++ DataDictionary.Resource/Enumerations/DomainPropertyType.Extension.cs
