--- DataDictionary.Resource/Enumerations/FileFormat.cs
@@ -1,2 +1,0 @@
-﻿// Place holder for file nesting purposes.
-namespace DataDictionary.Resource.Enumerations;
