+++ DataDictionary.Resource/Enumerations/FileFormatType.Enumeration.cs
