+++ DataDictionary.Resource/Enumerations/FileFormatType.Extension.cs
