--- DataDictionary.Resource/Enumerations/LibraryMember.cs
@@ -1,12 +1,0 @@
-﻿namespace DataDictionary.Resource.Enumerations;
-
-/// <summary>
-/// Interface for Database MemberType Key.
-/// </summary>
-public interface ILibraryMemberType
-{
-    /// <summary>
-    /// Type of Member (NameSpace, Type, Property, Field, ...)
-    /// </summary>
-    LibraryMemberType MemberType { get; }
-}
