+++ DataDictionary.Resource/Enumerations/LibraryMemberType.Enumeration.cs
