+++ DataDictionary.Resource/Enumerations/LibraryMemberType.Extension.cs
