+++ DataDictionary.Resource/Enumerations/NodeRenderAsType.Extension.cs
