--- DataDictionary.Resource/Enumerations/Scope.cs
@@ -1,12 +1,0 @@
-﻿namespace DataDictionary.Resource.Enumerations;
-
-/// <summary>
-/// Interface for the Scope
-/// </summary>
-public interface IScopeType
-{
-    /// <summary>
-    /// ScopeType for the Item.
-    /// </summary>
-    ScopeType Scope { get; }
-}
