+++ DataDictionary.Resource/Enumerations/ScopeType.Enumeration.cs
