+++ DataDictionary.Resource/Enumerations/ScopeType.Extension.cs
