--- DataDictionary.Resource/Enumerations/ScriptAs.cs
@@ -1,12 +1,0 @@
-﻿namespace DataDictionary.Resource.Enumerations;
-
-/// <summary>
-/// Interface for Scripting ScriptAs Key.
-/// </summary>
-public interface IScriptAsType
-{
-    /// <summary>
-    /// Type of Script that is Generated
-    /// </summary>
-    ScriptAsType ScriptAs { get; }
-}
