+++ DataDictionary.Resource/Enumerations/ScriptAsType.Enumeration.cs
