+++ DataDictionary.Resource/Enumerations/ScriptAsType.cs
@@ -1,6 +1,17 @@
 ﻿namespace DataDictionary.Resource.Enumerations
 {
     /// <summary>
+    /// Interface for Scripting ScriptAs Key.
+    /// </summary>
+    public interface IScriptAsType
+    {
+        /// <summary>
+        /// Type of Script that is Generated
+        /// </summary>
+        ScriptAsType ScriptAs { get; }
+    }
+
+    /// <summary>
     /// List of supported Script types.
     /// </summary>
     public enum ScriptAsType // TemplateScriptAsEnumeration
